Adjust the warning to send if _both_ of the tags are missing, not just _one_ of them
This is less strict - you only need to pass `repository` _or_
`repository_owner`. They do both need to be mapped in the ID pool, but
don't both need to be passed on to the chained role. One of them should
be, though, in the case of multiple orgs with the same repo. e.g.
`org-a/repo-a` and `org-b/repo-b`. You'll need to have the org name
available to IAM to verify this properly. This will be explained in docs
soon.