Support creating a `repository_name` claim when chaining (#4)
* Support creating a `repository_name` claim when chaining

GitHub doesn't provide a claim in its token, or an environment variable
/ context value containing just the repository's name. It does give the
owner in `repository_owner` though.

Having this available as a tag will be very useful when constructing
policies. For example, you could create a policy that allows repos to
push to s3 buckets named like

```
mycompany-${aws:RequestTag/repository_owner}-${aws:RequestTag/repository_name}
```

* README: Strengthen documentation around checking tags

Let's be more explicit about what people need to do when passing tags
with role chaining, and give an example for `repository_name` too.

* Update README: move paragraph down, update exmaple, add to TOC

* action: Add a warning if using `repository_name` without `repository` and `repository_owner`

These claims are needed to validate `repository_name` - so if they're
missing it implies maybe the user isn't doing the validation.

* Make the keys of our array the mapped claim name

They could be of the form `claim=tag`. So we need to split on the `=`
and do lookups on the first part. We still need the full thing around to
do the tag mapping.

* Adjust the warning to send if _both_ of the tags are missing, not just _one_ of them

This is less strict - you only need to pass `repository` _or_
`repository_owner`. They do both need to be mapped in the ID pool, but
don't both need to be passed on to the chained role. One of them should
be, though, in the case of multiple orgs with the same repo. e.g.
`org-a/repo-a` and `org-b/repo-b`. You'll need to have the org name
available to IAM to verify this properly. This will be explained in docs
soon.